[REFACT] dto 통일
