+++ src/main/java/com/programmers/smrtstore/domain/cart/presentation/dto/req/UpdateCartOptionRequest.java
@@ -1,4 +1,4 @@
-package com.programmers.smrtstore.domain.cart.application.dto.req;
+package com.programmers.smrtstore.domain.cart.presentation.dto.req;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -11,7 +11,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class UpdateCartOptionRequest {
-    private Long cartId;
-    private Long userId;
+
     private Long productDetailOptionId;
 }
